refactor: Rename `lib/plugins/aws/utils/getS3EndpointForRegion.js`
