+++ lib/plugins/aws/deploy/lib/validateTemplate.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const getS3EndpointForRegion = require('../../utils/getS3EndpointForRegion');
+const getS3EndpointForRegion = require('../../utils/get-s3-endpoint-for-region');
 const ServerlessError = require('../../../../serverless-error');
 
 module.exports = {
