+++ lib/plugins/aws/lib/update-stack.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const getS3EndpointForRegion = require('../utils/getS3EndpointForRegion');
+const getS3EndpointForRegion = require('../utils/get-s3-endpoint-for-region');
 const { log, progress } = require('@serverless/utils/log');
 const isChangeSetWithoutChanges = require('../utils/is-change-set-without-changes');
