+++ lib/plugins/aws/provider.js
@@ -5,7 +5,7 @@ const BbPromise = require('bluebird');
 const _ = require('lodash');
 const naming = require('./lib/naming.js');
 const fsp = require('fs').promises;
-const getS3EndpointForRegion = require('./utils/getS3EndpointForRegion');
+const getS3EndpointForRegion = require('./utils/get-s3-endpoint-for-region');
 const memoizeeMethods = require('memoizee/methods');
 const readline = require('readline');
 const { ALB_LISTENER_REGEXP } = require('./package/compile/events/alb/lib/validate');
