+++ lib/plugins/aws/utils/get-s3-endpoint-for-region.js
